refactor(kubernetes): extract common agent/apply impl
This extracted code will be consumed by a future `KubernetesService`
agent implementation. As these entities will both share a property
called `kubernetes_object_yaml` but are distinct types, this couldn't be
"hardcoded" into `agent_apply` and so the `stdin` extraction is left as
a macro which optionally takes a property field to override what is
expected to be a reasonably shared default.
